Display images alongside supplements and change QR-Code on click

+++ app/src/main/java/org/walleth/liquidaccomodationsupplements/model/liquiditynetwork/InvoiceRequest.kt
@@ -1,3 +1,3 @@
 package org.walleth.liquidaccomodationsupplements.model.liquiditynetwork
 
-data class InvoiceRequest(val amount: Int, val recipient: String)
+data class InvoiceRequest(val amount: Long, val recipient: String)
